Refactor 폴더 정리
