+++ src/Constants/constants.tsx
@@ -1,0 +1,3 @@
+export const PRIMARY_COLOR_SKY = "#b7c9e2";
+export const PRIMARY_COLOR_WHITE = "#f8f5f1";
+export const PRIMARY_COLOR_BLUE = "#7ba1da";
