+++ src/SignInUpPage/SignInUpBox.tsx
@@ -1,35 +1,44 @@
 //로그인 페이지
 import React, { useState } from "react";
-import { PRIMARY_COLOR_BLUE} from "../../src/constants";
+import { PRIMARY_COLOR_BLUE } from "../Constants/constants";
 // @ts-ignore
 import logo from "./logo.png";
 
 import styled from "styled-components";
 import LoginForm from "./LoginForm";
 import SignUpForm from "./SignUpForm";
-const SignInUpBox=()=>{
+const SignInUpBox = () => {
   const [isLoginTab, setIsLoginTab] = useState(true);
 
   const handleLoginSubmit = (username: string, password: string) => {
-    console.log('로그인', username, password);
+    console.log("로그인", username, password);
   };
 
-  const handleSignUpSubmit = (username: string, password: string, email: string, nickname: string) => {
-    console.log('회원가입', username, password, email, nickname);
+  const handleSignUpSubmit = (
+    username: string,
+    password: string,
+    email: string,
+    nickname: string
+  ) => {
+    console.log("회원가입", username, password, email, nickname);
   };
 
-    return(
-      <LogInBoxContainer>
+  return (
+    <LogInBoxContainer>
       <ContainerWithLogo>
         <Logo />
         <TabsContainer>
-        <Tabs>
-          <Tab onClick={() => setIsLoginTab(true)} active={isLoginTab}>Log In</Tab>
-          <Tab onClick={() => setIsLoginTab(false)} active={!isLoginTab}>Sign Up</Tab>
-        </Tabs>
-      </TabsContainer>
+          <Tabs>
+            <Tab onClick={() => setIsLoginTab(true)} active={isLoginTab}>
+              Log In
+            </Tab>
+            <Tab onClick={() => setIsLoginTab(false)} active={!isLoginTab}>
+              Sign Up
+            </Tab>
+          </Tabs>
+        </TabsContainer>
         <LogInBox>
-        <Contents visible={isLoginTab}>
+          <Contents visible={isLoginTab}>
             <LoginForm onSubmit={handleLoginSubmit} />
           </Contents>
           <Contents visible={!isLoginTab}>
@@ -38,42 +47,40 @@ const SignInUpBox=()=>{
         </LogInBox>
       </ContainerWithLogo>
     </LogInBoxContainer>
-    );
-
-}
+  );
+};
 export default SignInUpBox;
 
-
 const LogInBoxContainer = styled.div`
-  display  : flex ;
+  display: flex;
   align-items: center;
   flex-direction: column;
 `;
 
 const ContainerWithLogo = styled.div`
-  text-align  :center ;
+  text-align: center;
   margin-top: 125px;
-  display:flex;
-  flex-direction:column;
-  align-items:center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 `;
 
 const LogInBox = styled.div`
-    display: flex;
-    flex-direction: column;
-    text-align: left;
-    width:500px;
-    height: 500px;
-    background-color: ${PRIMARY_COLOR_BLUE};
-    border-radius: 30px;
-    box-shadow: 0 0 15px rgba(0,0,0,0.4);
+  display: flex;
+  flex-direction: column;
+  text-align: left;
+  width: 500px;
+  height: 500px;
+  background-color: ${PRIMARY_COLOR_BLUE};
+  border-radius: 30px;
+  box-shadow: 0 0 15px rgba(0, 0, 0, 0.4);
 `;
 
 const Logo = styled.div`
-    background: url(${logo});
-    width:120px;
-  height:50px;
-  background-position:center;
+  background: url(${logo});
+  width: 120px;
+  height: 50px;
+  background-position: center;
   background-size: cover;
   background-repeat: no-repeat;
 `;
@@ -84,7 +91,7 @@ const TabsContainer = styled.div`
 `;
 
 const Tabs = styled.div`
-    display: flex;
+  display: flex;
   width: 100%;
   justify-content: center;
   margin-top: 15px;
@@ -94,8 +101,9 @@ const Tab = styled.div<{ active: boolean }>`
   cursor: pointer;
   padding: 15px 30px;
   font-size: 18px;
-  color: ${props => props.active ? PRIMARY_COLOR_BLUE : '#ccc'};
-  border-bottom: ${props => props.active ? `2px solid ${PRIMARY_COLOR_BLUE}` : '2px solid transparent'};
+  color: ${(props) => (props.active ? PRIMARY_COLOR_BLUE : "#ccc")};
+  border-bottom: ${(props) =>
+    props.active ? `2px solid ${PRIMARY_COLOR_BLUE}` : "2px solid transparent"};
   transition: color 0.5s, border-bottom-color 0.5s;
   &:hover {
     color: black;
@@ -103,7 +111,7 @@ const Tab = styled.div<{ active: boolean }>`
 `;
 
 const Contents = styled.div<{ visible: boolean }>`
-   display: ${(props) => (props.visible ? "flex" : "none")};
+  display: ${(props) => (props.visible ? "flex" : "none")};
   justify-content: center;
   align-items: center;
   margin-top: 70px;
